+++ maintenance-api/src/main/java/eu/kennytv/maintenance/api/event/MaintenanceReloadedEvent.java
@@ -1,6 +1,6 @@
 /*
- * Maintenance - https://git.io/maintenancemode
- * Copyright (C) 2018-2021 KennyTV (https://github.com/KennyTV)
+ * This file is part of Maintenance - https://github.com/kennytv/Maintenance
+ * Copyright (C) 2018-2021 kennytv (https://github.com/kennytv)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 package eu.kennytv.maintenance.api.event;
 
 import eu.kennytv.maintenance.api.IMaintenance;
@@ -25,7 +24,7 @@ import eu.kennytv.maintenance.api.proxy.IMaintenanceProxy;
 /**
  * Notification event fired when Maintenance configuration is reloaded.
  *
- * @author KennyTV
+ * @author kennytv
  * @see IMaintenance#isMaintenance() for the new maintenance status
  * @see IMaintenanceProxy#getMaintenanceServers() for the new list of servers under maintenance
  * @since 3.0.6