+++ maintenance-api/src/main/java/eu/kennytv/maintenance/api/event/proxy/ServerMaintenanceChangedEvent.java
@@ -1,6 +1,6 @@
 /*
- * Maintenance - https://git.io/maintenancemode
- * Copyright (C) 2018-2021 KennyTV (https://github.com/KennyTV)
+ * This file is part of Maintenance - https://github.com/kennytv/Maintenance
+ * Copyright (C) 2018-2021 kennytv (https://github.com/kennytv)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 package eu.kennytv.maintenance.api.event.proxy;
 
 import eu.kennytv.maintenance.api.event.manager.MaintenanceEvent;
@@ -24,7 +23,7 @@ import eu.kennytv.maintenance.api.proxy.Server;
 /**
  * Notification event fired when maintenance mode has been changed on a proxied server.
  *
- * @author KennyTV
+ * @author kennytv
  * @since 3.0.1
  */
 public final class ServerMaintenanceChangedEvent implements MaintenanceEvent {