+++ maintenance-bungee/src/main/java/eu/kennytv/maintenance/bungee/listener/ProxyPingListener.java
@@ -1,6 +1,6 @@
 /*
- * Maintenance - https://git.io/maintenancemode
- * Copyright (C) 2018-2021 KennyTV (https://github.com/KennyTV)
+ * This file is part of Maintenance - https://github.com/kennytv/Maintenance
+ * Copyright (C) 2018-2021 kennytv (https://github.com/kennytv)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 package eu.kennytv.maintenance.bungee.listener;
 
 import eu.kennytv.maintenance.bungee.MaintenanceBungeePlugin;
@@ -36,7 +35,7 @@ public final class ProxyPingListener implements Listener {
 
     @EventHandler(priority = 80)
     public void proxyPing(final ProxyPingEvent event) {
-        if (!settings.isMaintenance()) return;
+        if (!settings.isMaintenance() || !settings.isEnablePingMessages()) return;
 
         final ServerPing ping = event.getResponse();
         ServerPing.Players players = ping.getPlayers();