+++ maintenance-core/src/main/java/eu/kennytv/maintenance/core/command/subcommand/DebugCommand.java
@@ -1,6 +1,6 @@
 /*
- * Maintenance - https://git.io/maintenancemode
- * Copyright (C) 2018-2021 KennyTV (https://github.com/KennyTV)
+ * This file is part of Maintenance - https://github.com/kennytv/Maintenance
+ * Copyright (C) 2018-2021 kennytv (https://github.com/kennytv)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 package eu.kennytv.maintenance.core.command.subcommand;
 
 import eu.kennytv.maintenance.core.MaintenancePlugin;
@@ -25,14 +24,13 @@ import eu.kennytv.maintenance.core.util.SenderInfo;
 public final class DebugCommand extends CommandInfo {
 
     public DebugCommand(final MaintenancePlugin plugin) {
-        super(plugin, "debug");
+        super(plugin, "debug", false);
     }
 
     @Override
     public void execute(final SenderInfo sender, final String[] args) {
         if (checkArgs(sender, args, 1)) return;
 
-        //TODO put this and the help message into language files
         plugin.setDebug(!plugin.isDebug());
         sender.sendMessage("Debug mode is now " + (plugin.isDebug() ? "enabled" : "disabled"));
     }