+++ maintenance-core/src/main/java/eu/kennytv/maintenance/core/command/subcommand/HelpCommand.java
@@ -1,6 +1,6 @@
 /*
- * Maintenance - https://git.io/maintenancemode
- * Copyright (C) 2018-2021 KennyTV (https://github.com/KennyTV)
+ * This file is part of Maintenance - https://github.com/kennytv/Maintenance
+ * Copyright (C) 2018-2021 kennytv (https://github.com/kennytv)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 package eu.kennytv.maintenance.core.command.subcommand;
 
 import com.google.common.base.Preconditions;
@@ -62,7 +61,7 @@ public final class HelpCommand extends CommandInfo {
         Preconditions.checkArgument(page > 0);
         final List<String> commands = new ArrayList<>();
         for (final CommandInfo cmd : plugin.getCommandManager().getCommands()) {
-            if (cmd.hasPermission(sender)) {
+            if (cmd.isVisible() && cmd.hasPermission(sender)) {
                 commands.add(cmd.getHelpMessage());
             }
         }
@@ -84,7 +83,7 @@ public final class HelpCommand extends CommandInfo {
         if (page * COMMANDS_PER_PAGE < commands.size()) {
             sender.sendMessage(getMessage("helpNextPage").replace("%PAGE%", Integer.toString(page + 1)));
         } else {
-            sender.sendMessage("§8× §eVersion " + plugin.getVersion() + " §7by §bKennyTV");
+            sender.sendMessage("§8× §eVersion " + plugin.getVersion() + " §7by §bkennytv");
         }
 
         sender.sendMessage(header);