+++ maintenance-core/src/main/java/eu/kennytv/maintenance/core/listener/JoinListenerBase.java
@@ -1,6 +1,6 @@
 /*
- * Maintenance - https://git.io/maintenancemode
- * Copyright (C) 2018-2021 KennyTV (https://github.com/KennyTV)
+ * This file is part of Maintenance - https://github.com/kennytv/Maintenance
+ * Copyright (C) 2018-2021 kennytv (https://github.com/kennytv)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 package eu.kennytv.maintenance.core.listener;
 
 import eu.kennytv.maintenance.core.MaintenancePlugin;
@@ -29,7 +28,6 @@ import java.util.UUID;
 public abstract class JoinListenerBase {
     protected final MaintenancePlugin plugin;
     protected final Settings settings;
-    protected final UUID notifyUuid = new UUID(-6334418481592579467L, -4779835342378829761L);
     private final Set<UUID> notifiedPlayers = new HashSet<>();
 
     protected JoinListenerBase(final MaintenancePlugin plugin, final Settings settings) {
@@ -59,10 +57,6 @@ public abstract class JoinListenerBase {
     }
 
     protected void updateCheck(final SenderInfo sender) {
-        if (sender.getUuid().equals(notifyUuid)) {
-            sender.sendMessage("§6Maintenance §aVersion " + plugin.getVersion());
-            return;
-        }
         if (!settings.hasUpdateChecks()) return;
         if (!sender.hasPermission("maintenance.admin") || notifiedPlayers.contains(sender.getUuid())) return;
         plugin.async(() -> {