+++ maintenance-core/src/main/java/eu/kennytv/maintenance/core/runnable/MaintenanceRunnableBase.java
@@ -1,6 +1,6 @@
 /*
- * Maintenance - https://git.io/maintenancemode
- * Copyright (C) 2018-2021 KennyTV (https://github.com/KennyTV)
+ * This file is part of Maintenance - https://github.com/kennytv/Maintenance
+ * Copyright (C) 2018-2021 kennytv (https://github.com/kennytv)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 package eu.kennytv.maintenance.core.runnable;
 
 import eu.kennytv.maintenance.core.MaintenancePlugin;
@@ -49,23 +48,7 @@ public abstract class MaintenanceRunnableBase implements Runnable {
     }
 
     public String getTime() {
-        final int preHours = this.seconds / 60;
-        final int minutes = preHours % 60;
-        final int seconds = this.seconds % 60;
-
-        final StringBuilder buider = new StringBuilder();
-        append(buider, "hour", preHours / 60);
-        append(buider, "minute", minutes);
-        append(buider, "second", seconds);
-        return buider.toString();
-    }
-
-    private void append(final StringBuilder builder, final String timeUnit, final int time) {
-        if (time == 0) return;
-        if (builder.length() != 0) {
-            builder.append(' ');
-        }
-        builder.append(time).append(' ').append(settings.getMessage(time == 1 ? timeUnit : timeUnit + "s"));
+        return plugin.getFormattedTime(seconds);
     }
 
     public boolean shouldEnable() {