+++ maintenance-velocity/src/main/java/eu/kennytv/maintenance/velocity/MaintenanceVelocityPlugin.java
@@ -1,6 +1,6 @@
 /*
- * Maintenance - https://git.io/maintenancemode
- * Copyright (C) 2018-2021 KennyTV (https://github.com/KennyTV)
+ * This file is part of Maintenance - https://github.com/kennytv/Maintenance
+ * Copyright (C) 2018-2021 kennytv (https://github.com/kennytv)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
 package eu.kennytv.maintenance.velocity;
 
 import com.google.inject.Inject;
@@ -39,6 +38,8 @@ import eu.kennytv.maintenance.core.dump.PluginDump;
 import eu.kennytv.maintenance.core.hook.ServerListPlusHook;
 import eu.kennytv.maintenance.core.proxy.MaintenanceProxyPlugin;
 import eu.kennytv.maintenance.core.proxy.SettingsProxy;
+import eu.kennytv.maintenance.core.proxy.util.ProfileLookup;
+import eu.kennytv.maintenance.core.proxy.util.ProxyOfflineSenderInfo;
 import eu.kennytv.maintenance.core.util.MaintenanceVersion;
 import eu.kennytv.maintenance.core.util.SenderInfo;
 import eu.kennytv.maintenance.core.util.ServerType;
@@ -69,10 +70,10 @@ import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
 /**
- * @author KennyTV
+ * @author kennytv
  * @since 3.0
  */
-@Plugin(id = "maintenance", name = "Maintenance", version = MaintenanceVersion.VERSION, authors = "KennyTV",
+@Plugin(id = "maintenance", name = "Maintenance", version = MaintenanceVersion.VERSION, authors = "kennytv",
         description = "Enable maintenance mode with a custom maintenance motd and icon.", url = "https://forums.velocitypowered.com/t/maintenance/129",
         dependencies = @Dependency(id = "serverlistplus", optional = true))
 public final class MaintenanceVelocityPlugin extends MaintenanceProxyPlugin {
@@ -109,7 +110,9 @@ public final class MaintenanceVelocityPlugin extends MaintenanceProxyPlugin {
         // ServerListPlus integration
         server.getPluginManager().getPlugin("serverlistplus").ifPresent(slpContainer -> slpContainer.getInstance().ifPresent(serverListPlus -> {
             serverListPlusHook = new ServerListPlusHook(serverListPlus);
-            serverListPlusHook.setEnabled(!settingsProxy.isMaintenance());
+            if (settings.isEnablePingMessages()) {
+                serverListPlusHook.setEnabled(!settingsProxy.isMaintenance());
+            }
             logger.info("Enabled ServerListPlus integration!");
         }));
     }
@@ -127,14 +130,11 @@ public final class MaintenanceVelocityPlugin extends MaintenanceProxyPlugin {
 
     @Override
     public void sendUpdateNotification(final SenderInfo sender) {
-        final TextComponent tc1 = translate(getPrefix());
-        final TextComponent tc2 = translate("§cDownload it at: §6https://www.spigotmc.org/resources/maintenance.40699/");
-        final TextComponent click = translate(" §7§l§o(CLICK ME)");
-        click.clickEvent(ClickEvent.openUrl("https://www.spigotmc.org/resources/maintenance.40699/"));
-        click.hoverEvent(HoverEvent.showText(translate("§aDownload the latest version")));
-        tc1.append(tc2);
-        tc1.append(click);
-        ((VelocitySenderInfo) sender).sendMessage(tc1);
+        final TextComponent component = translate(getPrefix()).append(translate("§cDownload it at: §6https://www.spigotmc.org/resources/maintenance.40699/"));
+        final TextComponent clickText = translate(" §7§l§o(CLICK ME)")
+                .clickEvent(ClickEvent.openUrl("https://www.spigotmc.org/resources/maintenance.40699/"))
+                .hoverEvent(HoverEvent.showText(translate("§aDownload the latest version")));
+        ((VelocitySenderInfo) sender).sendMessage(component.append(clickText));
     }
 
     public boolean isMaintenance(final RegisteredServer serverInfo) {
@@ -210,7 +210,18 @@ public final class MaintenanceVelocityPlugin extends MaintenanceProxyPlugin {
     @Nullable
     public SenderInfo getOfflinePlayer(final String name) {
         final Optional<Player> player = server.getPlayer(name);
-        return player.map(VelocitySenderInfo::new).orElse(null);
+        if (player.isPresent()) {
+            return new VelocitySenderInfo(player.get());
+        }
+
+        final ProfileLookup profile;
+        try {
+            profile = doUUIDLookup(name);
+        } catch (final IOException e) {
+            e.printStackTrace();
+            return null;
+        }
+        return new ProxyOfflineSenderInfo(profile.getUuid(), profile.getName());
     }
 
     @Override
@@ -231,6 +242,11 @@ public final class MaintenanceVelocityPlugin extends MaintenanceProxyPlugin {
     @Override
     public void async(final Runnable runnable) {
         server.getScheduler().buildTask(this, runnable).schedule();
+    }
+
+    @Override
+    protected void executeConsoleCommand(final String command) {
+        server.getCommandManager().executeAsync(server.getConsoleCommandSource(), command);
     }
 
     @Override